fix build error --- error[E0308]: mismatched types   --> C:\Users\Administrator\.cargo\registry\src\github.com-1ecc6299db9ec823\quickjs_runtime-0.8.0\src\quickjs_utils\typedarrays.rs:94:9    | 94 |         length,    |         ^^^^^^ expected `u32`, found `u64`    | help: you can convert a `u64` to a `u32` and panic if the converted value doesn't fit    | 94 |         length.try_into().unwrap(),    |               ++++++++++++++++++++
error[E0308]: mismatched types
   --> C:\Users\Administrator\.cargo\registry\src\github.com-1ecc6299db9ec823\quickjs_runtime-0.8.0\src\quickjs_utils\typedarrays.rs:157:59
    |
157 |     let raw = q::JS_NewArrayBufferCopy(ctx, buf.as_ptr(), length);
    |                                                           ^^^^^^ expected `u32`, found `u64`
    |
help: you can convert a `u64` to a `u32` and panic if the converted value doesn't fit
    |
157 |     let raw = q::JS_NewArrayBufferCopy(ctx, buf.as_ptr(), length.try_into().unwrap());
    |                                                                 ++++++++++++++++++++

error[E0308]: mismatched types
   --> C:\Users\Administrator\.cargo\registry\src\github.com-1ecc6299db9ec823\quickjs_runtime-0.8.0\src\quickjs_utils\typedarrays.rs:204:45
    |
204 |         let ptr = q::JS_GetArrayBuffer(ctx, &mut len, *array_buffer.borrow_value());
    |                                             ^^^^^^^^ expected `u32`, found `u64`
    |
    = note:    expected raw pointer `*mut u32`
            found mutable reference `&mut u64`

error[E0308]: mismatched types
   --> C:\Users\Administrator\.cargo\registry\src\github.com-1ecc6299db9ec823\quickjs_runtime-0.8.0\src\quickjs_utils\typedarrays.rs:236:41
    |
236 |     let ptr = q::JS_GetArrayBuffer(ctx, &mut len, *array_buffer.borrow_value());
    |                                         ^^^^^^^^ expected `u32`, found `u64`
    |
    = note:    expected raw pointer `*mut u32`
            found mutable reference `&mut u64`

+++ src/quickjs_utils/typedarrays.rs
@@ -91,7 +91,7 @@ pub unsafe fn new_array_buffer(
     let raw = q::JS_NewArrayBuffer(
         ctx,
         buffer_ptr,
-        length,
+        length as u32,
         Some(free_func),
         opaque as _,
         is_shared as _,
@@ -153,7 +153,7 @@ pub unsafe fn new_array_buffer_copy(
     ctx: *mut q::JSContext,
     buf: &[u8],
 ) -> Result<JSValueRef, JsError> {
-    let length = buf.len() as u64;
+    let length = buf.len() as u32;
     let raw = q::JS_NewArrayBufferCopy(ctx, buf.as_ptr(), length);
     let obj_ref = JSValueRef::new(
         ctx,
@@ -200,7 +200,7 @@ pub unsafe fn detach_array_buffer_buffer(
             buffers.remove(&id)
         })
     } else {
-        let mut len: u64 = 0;
+        let mut len: u32 = 0;
         let ptr = q::JS_GetArrayBuffer(ctx, &mut len, *array_buffer.borrow_value());
 
         Vec::from_raw_parts(ptr, len as usize, len as usize)
@@ -232,7 +232,7 @@ pub unsafe fn get_array_buffer_buffer_copy(
 ) -> Result<Vec<u8>, JsError> {
     debug_assert!(is_array_buffer(ctx, array_buffer));
 
-    let mut len: u64 = 0;
+    let mut len: u32 = 0;
     let ptr = q::JS_GetArrayBuffer(ctx, &mut len, *array_buffer.borrow_value());
 
     let slice = std::slice::from_raw_parts(ptr, len as usize);
